Cleanup code

+++ src/main/java/fr/envium/enviummod/addons/probe/theoneprobe/apiimpl/ProbeInfo.java
@@ -19,7 +19,7 @@ public class ProbeInfo extends ElementVertical {
     }
 
     public ProbeInfo() {
-        super((Integer) null, 2, ElementAlignment.ALIGN_TOPLEFT);
+        super(null, 2, ElementAlignment.ALIGN_TOPLEFT);
     }
 
     public static List<IElement> createElements(PacketBuffer buf) {
