+++ src/main/java/fr/envium/enviummod/core/blocks/EnviumChest.java
@@ -229,7 +229,7 @@ public class EnviumChest extends Block {
 	}
 
 	private static boolean isCatSittingOn(IWorld iWorld, BlockPos blockPos) {
-		List<CatEntity> list = iWorld.getEntitiesOfClass(CatEntity.class, new AxisAlignedBB((double) blockPos.getX(), (double) (blockPos.getY() + 1), (double) blockPos.getZ(), (double) (blockPos.getX() + 1), (double) (blockPos.getY() + 2), (double) (blockPos.getZ() + 1)));
+		List<CatEntity> list = iWorld.getEntitiesOfClass(CatEntity.class, new AxisAlignedBB(blockPos.getX(), blockPos.getY() + 1, blockPos.getZ(), blockPos.getX() + 1, blockPos.getY() + 2, blockPos.getZ() + 1));
 		if (!list.isEmpty()) {
 			for (CatEntity catentity : list) {
 				if (catentity.orderedToSit) {
